scripts/requirements.txt
azure-identity>=1.15.0
azure-ai-agents>=1.0.0
azure-ai-evaluation>=1.0.0
azure-search-documents>=11.4.0
openai>=1.12.0
python-dotenv>=1.0.0
pypdf>=4.0.0
pandas>=2.0.0
